backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
sqlalchemy==2.0.21
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
httpx==0.25.0
pydantic[email]==2.4.2
python-jose==3.3.0
email-validator==2.0.0
starlette==0.27.0
typing-extensions==4.7.1
cryptography==41.0.3
pyasn1==0.5.0
rsa==4.9
psycopg2-binary==2.9.9
alembic==1.12.0